Module blinkcast::static_mem

A channel implemented with static memory.

This module implements a broadcast channel with a fixed-size buffer, without allocation. The buffer (hosted by the Sender) is stored in static memory, it can be in the stack or in global static variables, and this can be done because Sender::new is a const fn.

When sending, we only need &Sender, so it can be done from multiple threads/cores as the same time.

The channel overwrites the oldest message if the buffer is full, prioritizing the latest data.

Key Features:

Broadcast: Multiple senders can send messages to multiple receivers simultaneously.

Fixed-size Buffer: Provides bounded memory usage with predictable performance.

Overwriting Behavior: Prioritizes the latest data in scenarios where the buffer becomes full.

Cloneable: the Receiver are cloneable, enabling flexible message distribution patterns, the Sender is not cloneable, but if setup in a global static variable, it can be used from multiple locations.

Usage Considerations:

Well-suited for scenarios where multiple components need to broadcast messages and the latest data takes priority.

Ideal when heap allocation is necessary or desirable.

Receivers must be fast enough to keep up with the senders and avoid losing messages due to overwriting.

Structs

Receiver

The receiver of the channel.

Sender

The sender of the channel.